Converted getUserByNick to return NULL on error, rather than false.
This function is defined to return a User*, however, rather than returning NULL
in the error case, it was returning false. Whilst strictly under C++98 this
is valid behaviour, it's rather obtuse, and a warning is emitted from clang
by default.

Rewriting this to return NULL as NULL in the error case fixes building with
clang++, since it's being passed -Wall.

Signed-off-by: Alastair Lynn <arplynn@gmail.com>