Added a LogType enumeration for the different types of logging levels. hanged occurrences of using the log via screen to use the logger class.
This change allows the logger class to be extended in the future to write
files for example

Removed the screen.log from plugin Api in favour of the new logger.log.
Corrected plugins on windows, without extern "C" on Windows MSVC mangles
the plugin methods and so the plugin loader can not find the init method.